fix live-updates for discussion-list (and make it a component)
