+++ app/components/discussion-list/template.html
@@ -1,5 +1,5 @@
 <div class="container">
-<div class="panel panel-default" ng-controller="DiscussionListCtrl">
+<div class="panel panel-default">
   <div class="panel-heading">
     <ul class="list-inline">
       <li>
