+++ app/controllers/index.ts
@@ -1,6 +1,5 @@
 import documentText from './documentText';
 import authReturn from './authReturn';
-import discussionList from './discussionList';
 import hivers from './hivers';
 import meta from './meta';
 import jobs from './jobs';
@@ -9,7 +8,6 @@ import passwordRequest from './passwordRequest';
 export default function(app) {
   documentText(app);
   authReturn(app);
-  discussionList(app);
   hivers(app);
   meta(app);
   jobs(app);
